CRUD Artikel Selesai

+++ admin/application/models/Article_model.php
@@ -10,24 +10,22 @@ class Article_model extends CI_Model {
     }
 
     public function get_data_where($where) {
-        $article = $this->mongo_db->where($where)->get($this->collection);
-        if (is_array($article) && count($article) > 0) {
-            $article = array_shift($article);
-            // $this->m$article['image']['tmpFilename'];
-            $data = $this->grid_fs->findOne(array('filename' => $article['image']['filename']));
-            header('Content-type: image/png;');
-            $stream = $data->getResource();
-
-            while (!feof($stream)) {
-                echo fread($stream, 8192);
-            }
-            die;
-        }
-        die;
+        return $this->mongo_db->where($where)->get($this->collection);
+    }
+    
+    public function get_image_filestream($where) {
+        return $this->grid_fs->findOne($where);
     }
 
-    public function upload_image_filestream($image, $options = []) {
+    public function upload_image_filestream($image, $options = [], $is_update = false) {
+        if ($is_update != false) {
+            $this->remove_image_filestream($options);
+        }
         return $this->grid_fs->storeUpload($image, $options);
+    }
+
+    public function remove_image_filestream($where) {
+        return $this->grid_fs->remove($where);
     }
 
     public function add_new_data($data) {
@@ -38,23 +36,32 @@ class Article_model extends CI_Model {
         return $this->mongo_db->insert($this->collection_detail, $data);
     }
 
-    public function update_user($data, $where) {
+    public function update_article($data, $where) {
         return $this->mongo_db->where($where)->set($data)->update($this->collection);
     }
 
-    public function remove_user($where) {
-        $this->mongo_db->where($where)->delete($this->collection);
+    public function remove_article($where) {
+        $article = $this->mongo_db->where($where)->get($this->collection);
+        if (is_array($article) && count($article) > 0) {
+            if (isset($artile['image'])) {
+                $this->remove_image_filestream(array('filename' => $article['image']['filename']));
+            }
+            unset($article);
+            $this->mongo_db->where($where)->delete($this->collection);
+        }
+        return true;
     }
 
-    public function remove_batch_user($data_where, $key) {
+    public function remove_batch_article($data_where, $key) {
         // $key => attr di database
         foreach ($data_where as $d_v) {
             $where[$key] = $d_v;
             if ($key == '_id') {
                 $where[$key] = new MongoId($d_v);
             }
-            $this->mongo_db->where($where)->delete($this->collection);
+            $this->remove_article($where);
         }
+        return true;
     }
 
 }
